feat(movie-controller): Add MovieController
Implement post, put and get requests in "/movies" path.

Implement getAllMovies and search in the same get controller, becase its
endpoints might cause a conflict.

Delete getAllMovies method in MovieService, because it is redundant with
search method and its filter values setted as null. It is mean that if
title is null, genre is null, and so on, beacause there is not any
request parameter, then search controller returns all the movies in
movie's repository.